Create binary tree node function
Added a function 'binary_tree_node' in the 'binary_trees.h' file to create a binary tree node. This function takes a parent node and a value, allocates memory for the new node, initializes its fields, and returns a pointer to the new node. In case of memory allocation failure, it returns NULL.