In-order traversal of a binary tree
Implemented 'binary_tree_inorder' function in 'binary_trees.h' for performing an in-order traversal of a binary tree. This function calls a provided function on each node, starting from the left subtree, then the root, and finally the right subtree. It checks that the tree and function pointers are not NULL to avoid errors.