Measure the depth of a node in a binary tree
Implemented 'binary_tree_depth' function in 'binary_trees.h' to measure the depth of a node in a binary tree. This function iteratively traverses the tree from the given node to its root (parent), counting the number of steps taken as the depth of the node. It returns 0 if the tree is NULL.