Count nodes with at least 1 child in a binary tree
Implemented 'binary_tree_nodes' function in 'binary_trees.h' to count the number of nodes in a binary tree that have at least 1 child. This function uses recursion and checks if a node is not NULL and has at least one child. It returns 0 if the tree is NULL or if it's a leaf node.