Measure binary tree height and balance factor
Added functions 'binary_tree_height' and 'binary_tree_balance' to 'binary_trees.h' to measure the height and balance factor of a binary tree, respectively. 'binary_tree_height' uses recursion to calculate the height of the tree, while 'binary_tree_balance' calculates the balance factor as the difference between the heights of the left and right subtrees.